requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
httpx==0.26.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# Utilities
python-dotenv==1.0.0
slugify==0.0.1
python-slugify==8.0.1

# Caching (optional but recommended)
redis==5.0.1
aioredis==2.0.1

# Monitoring and logging
loguru==0.7.2

# CORS
fastapi-cors==0.0.6